refactor(server/events): extract makeAwarenessObserver + add BufferedSelection type
Add BufferedSelection to types.ts (canonical shared type for awareness↔chat
selection buffer). Move awareness observer + getDwellMs private helper into
observers/awareness.ts. queue.ts selectionBuffer now typed as
Map<string, BufferedSelection>; getDwellMs and related constants removed.
