+++ src/server/events/types.ts
@@ -9,3 +9,14 @@
  */
 
 export * from "../../shared/events/types.js";
+
+/**
+ * Per-document selection buffer entry. Produced by the awareness observer
+ * after a dwell timeout and consumed by the ctrl-chat observer when attaching
+ * selection context to a chat message.
+ */
+export type BufferedSelection = {
+  from: number; // FlatOffset
+  to: number; // FlatOffset
+  selectedText: string;
+};
